pyproject.toml
[tool.poetry]
name = "academic-journal-reviewer"
version = "0.1.0"
description = "Agentic AI system for academic journal review"
authors = ["Muhammad Abiodun SULAIMAN"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
motor = "^3.3.2"
pymongo = "^4.6.0"
python-multipart = "^0.0.18"
pydantic = "^2.5.0"
python-docx = "^1.1.0"
PyPDF2 = "^3.0.1"
google-generativeai = "^0.3.2"
groq = "^0.4.1"
openai = "^1.3.0"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
autoflake = "^2.2.1"
isort = "^5.12.0"

[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
/(
    \.venv
  | venv
  | node_modules
  | __pycache__
  | \.git
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "venv", "node_modules", "__pycache__", ".git", "build", "dist"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
